+++ packages/@tinacms/cli/src/next/commands/audit-command/audit.ts
@@ -148,14 +148,12 @@ export const auditDocuments = async (args: AuditArgs) => {
             }
           })
       }
-      const params = transformDocumentIntoMutationRequestPayload(
+      const tinaSchema = await database.getSchema()
+      const values = mergeValuesWithDefaults(
         docResult.data.document._values,
-        {
-          includeCollection: true,
-          includeTemplate: typeof collection.templates !== 'undefined',
-        },
         topLevelDefaults
       )
+      const params = tinaSchema.transformPayload(collection.name, values)
 
       const mutation = `mutation($collection: String!, $relativePath: String!, $params: DocumentUpdateMutation!) {
         updateDocument(
@@ -188,64 +186,11 @@ export const auditDocuments = async (args: AuditArgs) => {
   return error
 }
 
-// TODO: move this to its own package
-export const transformDocumentIntoMutationRequestPayload = (
-  document: {
-    _collection: string
-    __typename?: string
-    _template: string
-    [key: string]: unknown
-  },
-  /** Whether to include the collection and template names as top-level keys in the payload */
-  instructions: { includeCollection?: boolean; includeTemplate?: boolean },
-  defaults?: any
+export const mergeValuesWithDefaults = (
+  document: object,
+  defaults?: object
 ) => {
-  const { _collection, __typename, _template, ...rest } = document
-
-  const params = transformParams(rest)
-  const paramsWithTemplate = instructions.includeTemplate
-    ? { [_template]: params }
-    : params
-
-  return instructions.includeCollection
-    ? { [_collection]: { ...defaults, ...filterObject(paramsWithTemplate) } }
-    : { ...defaults, ...filterObject(paramsWithTemplate) }
-}
-
-const transformParams = (data: unknown) => {
-  if (['string', 'number', 'boolean'].includes(typeof data)) {
-    return data
-  }
-  if (Array.isArray(data)) {
-    return data.map((item) => transformParams(item))
-  }
-  try {
-    assertShape<{ _template: string; __typename?: string }>(data, (yup) =>
-      // @ts-ignore No idea what yup is trying to tell me:  Type 'RequiredStringSchema<string, Record<string, any>>' is not assignable to type 'AnySchema<any, any, any>
-      yup.object({ _template: yup.string().required() })
-    )
-    const { _template, __typename, ...rest } = data
-    const nested = transformParams(rest)
-    return { [_template]: nested }
-  } catch (e) {
-    if (e.message === 'Failed to assertShape - _template is a required field') {
-      if (!data) {
-        return undefined
-        return []
-      }
-      const accum = {}
-      Object.entries(data).map(([keyName, value]) => {
-        accum[keyName] = transformParams(value)
-      })
-      return accum
-    } else {
-      if (!data) {
-        return undefined
-        return []
-      }
-      throw e
-    }
-  }
+  return { ...defaults, ...filterObject(document) }
 }
 
 // SRC: https://stackoverflow.com/questions/39977214/merge-in-es6-es7object-assign-without-overriding-undefined-values