+++ packages/@tinacms/cli/src/next/commands/audit-command/index.ts
@@ -25,6 +25,10 @@ export class AuditCommand extends Command {
   noTelemetry = Option.Boolean('--noTelemetry', false, {
     description: 'Disable anonymous telemetry that is collected',
   })
+  datalayerPort = Option.String('--datalayer-port', '9000', {
+    description:
+      'Specify a port to run the datalayer server on. (default 9000)',
+  })
   static usage = Command.Usage({
     category: `Commands`,
     description: `Audit config and content files`,
@@ -39,7 +43,7 @@ export class AuditCommand extends Command {
   }
 
   async execute(): Promise<number | void> {
-    const configManager = new ConfigManager(this.rootPath)
+    const configManager = new ConfigManager({ rootPath: this.rootPath })
     logger.info('Starting Tina Audit')
 
     try {
@@ -53,9 +57,10 @@ export class AuditCommand extends Command {
     }
 
     // Initialize the host TCP server
-    createDBServer()
+    createDBServer(Number(this.datalayerPort))
     const database = await createAndInitializeDatabase(
       configManager,
+      Number(this.datalayerPort),
       this.clean ? undefined : new AuditFileSystemBridge(configManager.rootPath)
     )
     const { tinaSchema, graphQLSchema } = await buildSchema(