+++ packages/@tinacms/cli/src/next/commands/init-command/index.ts
@@ -4,8 +4,9 @@ import { initStaticTina } from '../../../cmds/init'
 
 export class InitCommand extends Command {
   static paths = [['init']]
-  rootPath = Option.String('--rootPath', {
-    description: 'Specify the root directory to run the CLI from',
+  pathToForestryConfig = Option.String('--forestryPath', {
+    description:
+      'Specify the relative path to the .forestry directory, if importing an existing forestry site.',
   })
   noTelemetry = Option.Boolean('--noTelemetry', false, {
     description: 'Disable anonymous telemetry that is collected',
@@ -23,7 +24,7 @@ export class InitCommand extends Command {
 
   async execute(): Promise<number | void> {
     await initStaticTina({
-      rootPath: this.rootPath || process.cwd(),
+      pathToForestryConfig: this.pathToForestryConfig || process.cwd(),
       noTelemetry: this.noTelemetry,
     })
     process.exit()