+++ packages/@tinacms/toolkit/src/plugins/branch-switcher/BranchSwitcher.tsx
@@ -122,6 +122,7 @@ export const BranchSwitcher = ({
   // Keep branch list up to date
   React.useEffect(() => {
     if (listState === 'ready') {
+      const cancelFuncs = []
       // update all branches that have indexing status of 'inprogress' or 'unknown'
       branchList
         .filter(
@@ -130,16 +131,35 @@ export const BranchSwitcher = ({
             x?.indexStatus?.status === 'unknown'
         )
         .forEach(async (x) => {
-          const indexStatus = await cms.api.tina.waitForIndexStatus({
+          const [
+            // When this promise resolves, we know the index status is no longer 'inprogress' or 'unknown'
+            waitForIndexStatusPromise,
+            // Calling this function will cancel the polling
+            cancelWaitForIndexFunc,
+          ] = cms.api.tina.waitForIndexStatus({
             ref: x.name,
           })
-          setBranchList((prev) => {
-            const newList = Array.from(prev)
-            const index = newList.findIndex((y) => y.name === x.name)
-            newList[index].indexStatus = indexStatus
-            return newList
-          })
+          cancelFuncs.push(cancelWaitForIndexFunc)
+          waitForIndexStatusPromise
+            .then((indexStatus) => {
+              setBranchList((previousBranchList) => {
+                // update the index status of the branch
+                const newBranchList = Array.from(previousBranchList)
+                const index = newBranchList.findIndex((y) => y.name === x.name)
+                newBranchList[index].indexStatus = indexStatus
+                return newBranchList
+              })
+            })
+            .catch((e) => {
+              if (e.message === 'AsyncPoller: cancelled') return
+              console.error(e)
+            })
         })
+      return () => {
+        cancelFuncs.forEach((x) => {
+          x()
+        })
+      }
     }
   }, [listState, branchList.length])